BUGID 3989: save custom field values only to db if they are not empty

+++ lib/functions/cfield_mgr.class.php
@@ -7,7 +7,7 @@
  * @author 		franciscom
  * @copyright 	2005-2009, TestLink community
  * @copyright 	Mantis BT team (some parts of code was reuse from the Mantis project) 
- * @version    	CVS: $Id: cfield_mgr.class.php,v 1.96.2.1 2010/11/09 11:11:09 asimon83 Exp $
+ * @version    	CVS: $Id: cfield_mgr.class.php,v 1.96.2.2 2010/11/09 14:38:00 mx-julian Exp $
  * @link 		http://www.teamst.org/index.php
  *
  * @internal Revisions:
@@ -862,6 +862,7 @@ function _get_ui_mgtm_cfg_for_node_type($map_node_id_cfg)
           $sql = "/* $debugMsg */ UPDATE {$this->tables['cfield_design_values']} " .
                  " SET value='{$safe_value}' " .
     	         " WHERE field_id={$field_id} AND	node_id={$node_id}";
+          $this->db->exec_query($sql);
         }
         // BUGID 3989
         else if ($value != "")
@@ -873,12 +874,14 @@ function _get_ui_mgtm_cfg_for_node_type($map_node_id_cfg)
   		    $sql = "/* $debugMsg */ INSERT INTO {$this->tables['cfield_design_values']} " .
   				   " ( field_id, node_id, value ) " .
   				   " VALUES	( {$field_id}, {$node_id}, '{$safe_value}' )";
+  		    $this->db->exec_query($sql);
   		// BUGID 3989
         } else {
   			$sql = "/* $debugMsg */ DELETE FROM {$this->tables['cfield_design_values']} " .
   				   " WHERE field_id={$field_id} AND	node_id={$node_id}";
+  			$this->db->exec_query($sql);
   		}
-        $this->db->exec_query($sql);
+  		
       } //foreach($cfield
     } //if( !is_null($cfield) )
 
@@ -1667,10 +1670,10 @@ function name_is_unique($id,$name)
           $sql = "UPDATE {$this->tables['cfield_execution_values']} " .
                  " SET value='{$safe_value}' " .
     	         $where_clause;
-    	    // file_put_contents('c:\update.txt',$sql);         
+    	  $this->db->exec_query($sql);      
         }
         // BUGID 3989
-        else if ($value != "")
+        else if (count($rs) == 0 && $value != "")
         {
 
           # Remark got from Mantis code:
@@ -1680,15 +1683,15 @@ function name_is_unique($id,$name)
   		  $sql = "INSERT INTO {$this->tables['cfield_execution_values']} " .
   				 " ( field_id, tcversion_id, execution_id,testplan_id,value ) " .
   			     " VALUES	( {$field_id}, {$node_id}, {$execution_id}, {$testplan_id}, '{$safe_value}' )";
-          // file_put_contents('c:\insert.txt',$sql);  
+		  $this->db->exec_query($sql);
         
   		// BUGID 3989
-        } else {
+        } else if (count($rs) > 0 && $value == "") {
   			$sql = "/* $debugMsg */ DELETE FROM {$this->tables['cfield_execution_values']} " .
-  				   " WHERE field_id={$field_id} AND	node_id={$node_id}";
+  				   $where_clause;
+  			$this->db->exec_query($sql);
   		}
-                               
-        $this->db->exec_query($sql);
+        
       } //foreach($cfield
     } //if( !is_null($cfield) )
           // file_put_contents('c:\bye.txt','bye');
@@ -2120,9 +2123,10 @@ function getXMLServerParams($node_id)
 	      $sql = "UPDATE {$this->tables['cfield_testplan_design_values']} " .
 	             " SET value='{$safe_value}' " .
 			     " WHERE field_id={$field_id} AND	link_id={$link_id}";
+	      $this->db->exec_query($sql);
 	    }
 	    // BUGID 3989
-	    else if ($value != "")
+	    else if ($this->db->num_rows( $result ) == 0 && $value != "")
 	    {
 	      # Remark got from Mantis code:
 		    # Always store the value, even if it's the dafault value
@@ -2131,12 +2135,14 @@ function getXMLServerParams($node_id)
 		    $sql = "INSERT INTO {$this->tables['cfield_testplan_design_values']} " .
 				   " ( field_id, link_id, value ) " .
 				   " VALUES	( {$field_id}, {$link_id}, '{$safe_value}' )";
+		    $this->db->exec_query($sql);
 	    // BUGID 3989
-        } else {
+        } else if ($this->db->num_rows( $result ) > 0 && $value == "") {
   			$sql = "/* $debugMsg */ DELETE FROM {$this->tables['cfield_testplan_design_values']} " .
-  				   " WHERE field_id={$field_id} AND	node_id={$node_id}";
+  				   " WHERE field_id={$field_id} AND	link_id={$link_id}";
+  			$this->db->exec_query($sql);
   		}
-	    $this->db->exec_query($sql);
+
 	  } //foreach($cfield
 	} //if( !is_null($cfield) )
